Downloader compatability
Moving creditclaim.py to its own folder and changing the camelcase name.
